requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
passlib==1.7.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# EPG & XML
xmltodict==0.13.0
defusedxml==0.7.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
unidecode==1.3.7

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Async utilities
asyncio-throttle==1.0.2

# Logging & Monitoring
colorlog==6.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
jinja2==3.1.2
